Refactor, fix timer code
