+++ leds.c
@@ -1,0 +1,13 @@
+#include "leds.h"
+
+// function to test board; red and white LEDs should alternate every second
+void leds() {
+    uint32_t last_millis = millis();
+    while(1) {
+        if (millis() - 1000 > last_millis){
+            last_millis = millis();
+            RED = RED == 1 ? 0 : 1;
+            WHITE = WHITE == 1 ? 0 : 1;
+        }
+    }
+}
